+++ CHANGELOG.md
@@ -9,4 +9,5 @@
  - Add Changelog [#107](https://github.com/AmolKumarGupta/Websnapper/issues/107)
 
 ### Fixes
+ - Add `thumbnail` config to handle thumbnail related things
  - Add Alert before leaving the page while recording [#106](https://github.com/AmolKumarGupta/Websnapper/issues/106)